Rename package "packages" to "commands"
Edit eventParser() in Parser.java

Add eventStartParser() and eventEndParser() to Parser.java to aid
eventParser() in parsing the event startDate, startTime, endDate and
endTime.
