+++ src/main/java/duke/commands/Command.java
@@ -1,4 +1,4 @@
-package duke.packages;
+package duke.commands;
 
 import java.util.ArrayList;
