+++ src/main/java/duke/commands/FindCommand.java
@@ -1,4 +1,4 @@
-package duke.packages;
+package duke.commands;
 
 import java.util.ArrayList;
