+++ src/main/java/duke/commands/ReminderCommand.java
@@ -1,4 +1,4 @@
-package duke.packages;
+package duke.commands;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
